Started adding code to identify and flag sensors, specifically REV ColorSensorV3 that do not initialize. But it still fails, and at various locations.
It appears that what is needed is a timeout on I2C (Wire) Read, but
that is not implemented in the SAMD21 version of TwoWire.

The recurring failure is on the CSV3 on the 2nd I2C bus on the
Climber.

For this commit, remove the CSV3's from the Climber configuration.

Added Wing_2025 to Climber and changed Feather on Climber as part of
this update also (added battery diode to another board and swapped).